Fixed cycle bug.

+++ core/cpu.cpp
@@ -732,7 +732,7 @@ void CPU::DecodeExecute()
 void CPU::ExecuteFrame( bool *isPaused ) // NOLINT
 {
     _isPaused = *isPaused;
-    u16 const currentFrame = _bus->ppu.GetFrame();
+    u64 const currentFrame = _bus->ppu.GetFrame();
     while ( currentFrame == _bus->ppu.GetFrame() ) {
         if ( _isPaused ) {
             break;
